Load/Save MIDI channel in FPS
git-svn-id: https://svn.code.sf.net/p/fsthost/code/trunk@8 75a15a69-0cfc-47d9-a1bd-ce2941b697b2

+++ jackvst.h
@@ -20,9 +20,9 @@ struct _JackVST {
     jack_port_t  *midi_outport;
     jack_port_t  **inports;
     jack_port_t  **outports;
-    int            bypassed;
-    int		   channel;
-    int            with_editor;
+    short          bypassed;
+    int            channel;
+    short          with_editor;
     double         tempo;
 
     /* For VST/i support */
